Add a macro to instantiate bivariate_sumcheck tests
