+++ crates/core/src/protocols/sumcheck/v3/bivariate_mlecheck.rs
@@ -517,45 +517,3 @@ fn calculate_round_evals<'a, F: TowerField, Hal: ComputeLayer<F>>(
 	let evals = TryInto::<[F; 2]>::try_into(evals).expect("kernel returns two values");
 	Ok(evals)
 }
-
-#[cfg(test)]
-mod tests {
-	use binius_compute::cpu::layer::CpuLayerHolder;
-	use binius_core_test_utils::bivariate_sumcheck::generic_test_bivariate_mlecheck_prove_verify;
-	use binius_fast_compute::layer::FastCpuLayerHolder;
-	use binius_field::{
-		arch::OptimalUnderlier, as_packed_field::PackedType, tower::CanonicalTowerFamily,
-	};
-	use binius_math::B128;
-
-	#[test]
-	fn test_bivariate_mlecheck_prove_verify() {
-		let compute_holder = CpuLayerHolder::<B128>::new(1 << 13, 1 << 12);
-		let n_vars = 8;
-		let n_multilins = 8;
-		let n_compositions = 8;
-		generic_test_bivariate_mlecheck_prove_verify(
-			compute_holder,
-			n_vars,
-			n_multilins,
-			n_compositions,
-		);
-	}
-
-	#[test]
-	fn test_bivariate_mlecheck_prove_verify_fast() {
-		type F = B128;
-		type P = PackedType<OptimalUnderlier, F>;
-
-		let compute_holder = FastCpuLayerHolder::<CanonicalTowerFamily, P>::new(1 << 13, 1 << 12);
-		let n_vars = 8;
-		let n_multilins = 8;
-		let n_compositions = 8;
-		generic_test_bivariate_mlecheck_prove_verify(
-			compute_holder,
-			n_vars,
-			n_multilins,
-			n_compositions,
-		);
-	}
-}
