+++ crates/core/src/protocols/sumcheck/v3/bivariate_product.rs
@@ -429,69 +429,3 @@ pub enum PhaseState<F: Field> {
 	InitialSums(Vec<F>),
 	BatchedSum(F),
 }
-
-#[cfg(test)]
-mod tests {
-	use binius_compute::cpu::layer::CpuLayerHolder;
-	use binius_core_test_utils::bivariate_sumcheck::{
-		generic_test_bivariate_sumcheck_prove_verify, generic_test_calculate_round_evals,
-	};
-	use binius_fast_compute::layer::FastCpuLayerHolder;
-	use binius_field::{
-		BinaryField128b, arch::OptimalUnderlier, as_packed_field::PackedType,
-		tower::CanonicalTowerFamily,
-	};
-	use binius_math::B128;
-
-	#[test]
-	fn test_calculate_round_evals() {
-		let compute_holder = CpuLayerHolder::new(1 << 11, 1 << 10);
-		let n_vars = 8;
-		generic_test_calculate_round_evals(compute_holder, n_vars)
-	}
-
-	#[test]
-	fn test_calculate_round_evals_fast_cpu() {
-		type F = BinaryField128b;
-		type Packed = PackedType<OptimalUnderlier, F>;
-
-		let compute_holder =
-			FastCpuLayerHolder::<CanonicalTowerFamily, Packed>::new(1 << 11, 1 << 10);
-		let n_vars = 8;
-		generic_test_calculate_round_evals(compute_holder, n_vars)
-	}
-
-	#[test]
-	fn test_bivariate_sumcheck_prove_verify() {
-		let n_vars = 8;
-		let n_multilins = 8;
-		let n_compositions = 8;
-
-		let compute_holder = CpuLayerHolder::<B128>::new(1 << 13, 1 << 12);
-		generic_test_bivariate_sumcheck_prove_verify(
-			compute_holder,
-			n_vars,
-			n_multilins,
-			n_compositions,
-		)
-	}
-
-	#[test]
-	fn test_bivariate_sumcheck_prove_verify_fast() {
-		type F = BinaryField128b;
-		type Packed = PackedType<OptimalUnderlier, F>;
-
-		let n_vars = 8;
-		let n_multilins = 8;
-		let n_compositions = 8;
-		let compute_holder =
-			FastCpuLayerHolder::<CanonicalTowerFamily, Packed>::new(1 << 13, 1 << 12);
-
-		generic_test_bivariate_sumcheck_prove_verify(
-			compute_holder,
-			n_vars,
-			n_multilins,
-			n_compositions,
-		)
-	}
-}
